+++ envelope/app/user_channel/apps.py
@@ -5,5 +5,6 @@ class UserChannelConfig(AppConfig):
     name = "envelope.app.user_channel"
 
     def ready(self):
-        from . import channel
-        from . import async_signals
+        from . import channel  # noqa
+        from . import async_signals  # noqa
+        from . import signals  # noqa