+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added setting `ENVELOPE_USER_CHANNEL_SEND_SUBSCRIBE` to send a subscribe message to consumer rather than just adding
   the client to the users channel.
+* `user_logged_out` signal causes a close message to be sent on the user channel, so consumers will be disconnected. 
 * Loading RQ-job via message class for deferred jobs, to make the code easier to follow + overrides simpler.
 
 ## 1.0.3 (2024-03-13)